Fix filename handling in metadata graph (#21)
* Change representation of filename in metadata graph

Switch from `<path/to/file>` to `<file://path/to/file>` because the latter can be generated by a sparql query, the former is not actually a valid URI and sparql does not let you make it. The latter is not really a valid file URI either (only two slashes), but it works.

* Add message and indexVar to ForeachStep

* Allow inline strings  in <values>

* Improve error/usage messages in <write>

* Improve info output in <until>

* Improve exception message in PipelineMojo

* Change plugin.xml version to next release version